Split out agreement from transaction
